Completed unit tests for patterns and channels.
Note that older unit tests instantiate real patterns.
I've added a (empty) TestPattern to be used for testing.
That's because we can't be certain other patterns will
be present when running tests because they will become
submodules, rather than part of the repository.

These real patterns used in unit tests will need to be replaced
with the TestPattern at some later stage.