Fixes #102
This was a tricky one. When using a key for a snippet that is the
same as a key for a point, the Designer theme will overwrite your
snippet with a debugPoint snippet.

That behaviour is expected, but we should still avoid clashing
snippet IDs. So I've made the Designer theme add some weird
prefix to the snippets it adds. This should avoid clashing snippet
IDs.