Redesigned the BruceBoxerBriefs pattern, fixes #7
This was the first pattern I ever designed. And while it's wildly
popular, it has some issues. This is a redesign with intro a more
controlled pattern that matches seamlengths.

I've also removed the vertical stretch option, and added other
options. A backRise options to rise the pattern at the back,
and a bulge option that does exactly what you think.

Thanks to Stefan for his valuable input on the issues with this pattern.