Fixed issue in Part::angle(), dealt with ripple effects of that issue
Turns out the angle() method returned results that were 180 degrees off.
This is kind of a big deal, because angle() is used in a lot of places,
so there will be a bunch of updates to fix them all.

While this bug should have never been there, I suspect it was introduced
when trying to work out what quadrant of a circle an angle sits in.
While doing that, it's likely that I got confused thinking that a positive
delta on the Y-axis puts the angle in the top quadrant (as common sense
would dictate) but that I lost track for a moment of the fact that in SVG
the coordinate system works differently, and the Y-axis increments downwards.

This also caused changes in the Part::rotate() method, and a bunch of other
small adjustments.

I've also rewrote the Part::shiftTowards() methos to be mode concise.