:sparkles: Back neck cutout and armhole depth tweaks in Brian
Back neck cutout was bolted down to 20mm. This is now configurable
although it is not added as an option (as that would be somewhat
disruptive to patterns extending Brian).
In Brian itself, it's still bolted down to 20mm, but unlike before,
changing the value will correctly adapt the pattern.

This also includes a chnge to make the armhole depth take
the shoulder slope into account, as it should.